Fix some code

+++ src/main/java/ru/skypro/ads/service/AdsService.java
@@ -1,9 +1,9 @@
 package ru.skypro.ads.service;
 
 import org.springframework.data.util.Pair;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Isolation;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.ads.component.RecordMapper;
 import ru.skypro.ads.dto.*;
@@ -11,10 +11,11 @@ import ru.skypro.ads.entity.Ads;
 import ru.skypro.ads.entity.Comment;
 import ru.skypro.ads.entity.Image;
 import ru.skypro.ads.entity.User;
+import ru.skypro.ads.exception.AdsNotFoundException;
+import ru.skypro.ads.exception.ImageNotFoundException;
 import ru.skypro.ads.repository.AdsRepository;
 import ru.skypro.ads.repository.CommentRepository;
 import ru.skypro.ads.repository.ImageRepository;
-import ru.skypro.ads.repository.UserRepository;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,19 +25,17 @@ public class AdsService {
 
     private final AdsRepository adsRepository;
     private final CommentRepository commentRepository;
-    private final UserRepository userRepository;
     private final RecordMapper recordMapper;
     private final ImageService imageService;
     private final ImageRepository imageRepository;
     private final UserService userService;
 
-    private AdsService(AdsRepository adsRepository,
+    public AdsService(AdsRepository adsRepository,
                        CommentRepository commentRepository,
-                       UserRepository userRepository, RecordMapper recordMapper, ImageService imageService,
+                       RecordMapper recordMapper, ImageService imageService,
                        ImageRepository imageRepository, UserService userService) {
         this.adsRepository = adsRepository;
         this.commentRepository = commentRepository;
-        this.userRepository = userRepository;
         this.recordMapper = recordMapper;
         this.imageService = imageService;
         this.imageRepository = imageRepository;
@@ -55,74 +54,81 @@ public class AdsService {
         return responseWrapperAds;
     }
 
-    public ResponseEntity<AdsRecord> addAds(CreateAdsReq createAdsReq, MultipartFile multipartFile) {
+    public AdsRecord addAds(CreateAdsReq createAdsReq, MultipartFile multipartFile) {
         Ads ads = recordMapper.toEntity(createAdsReq);
         System.out.println(ads);
         UserRecord user = userService.getUser();
         ads.setAuthor(recordMapper.toEntity(user));
-        //TODO придумать что-нибудь тут
         ads = adsRepository.save(ads);
         ads.setImage(imageService.addImage(ads, multipartFile));
         ads = adsRepository.save(ads);
-        return new ResponseEntity<>(recordMapper.toRecord(ads), HttpStatus.OK);
-
+        return recordMapper.toRecord(ads);
     }
 
-    public ResponseEntity<ResponseWrapperComment> getAllCommentsById(Integer id) {
+    public ResponseWrapperComment getAllCommentsById(Integer id) {
         List<CommentRecord> list = new ArrayList<>();
         List<Comment> listComments = commentRepository.findAllById(id);
 
         for (int i = 0; i < listComments.size(); i++) {
             list.add(i, recordMapper.toRecord(listComments.get(i)));
         }
-        return new ResponseEntity<>(new ResponseWrapperComment(list.size(), list), HttpStatus.OK);
+        return new ResponseWrapperComment(list.size(), list);
     }
 
-    public ResponseEntity<CommentRecord> addComments(Integer adsId, CommentRecord commentRecord) {
+    public CommentRecord addComments(Integer adsId, CommentRecord commentRecord) {
         Comment comment = recordMapper.toEntity(commentRecord);
-        comment.setAds(adsRepository.findAdsById(adsId));
+        comment.setAds(adsRepository.findById(adsId).orElseThrow(() -> new AdsNotFoundException()));
         commentRepository.save(comment);
-        return new ResponseEntity<>(commentRecord, HttpStatus.OK);
+        return commentRecord;
     }
 
-    public ResponseEntity<FullAdsRecord> getFullAds(Integer id) {
-        Ads ads = adsRepository.findAdsById(id);
-        return new ResponseEntity<>(recordMapper.toFullAdsRecord(ads), HttpStatus.OK);
+    public FullAdsRecord getFullAds(Integer id) {
+        Ads ads = adsRepository.findById(id).orElseThrow(() -> new AdsNotFoundException());
+        return recordMapper.toFullAdsRecord(ads);
     }
 
-    public ResponseEntity<Void> deleteAds(Integer id) {
-        adsRepository.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+    @Transactional(isolation = Isolation.READ_COMMITTED)
+    public void deleteAds(Integer id) {
+        //TODO написать удаление комментов и картинок
+        Ads ads = adsRepository.findById(id).orElseThrow(() -> new AdsNotFoundException());
+        Image image = ads.getImage();
+        if (image != null) {
+            if (imageService.deleteImage(image) && commentRepository.removeAllByAdsId(ads.getId())) {
+                adsRepository.deleteById(id);
+                //TODO Написать эксепшны и добавить в хендлер
+            } else throw new RuntimeException("Exception with delete image or comments");
+        } else throw new RuntimeException("Exception with find image");
+
+
     }
 
-    public ResponseEntity<AdsRecord> updateAds(Integer id, CreateAdsReq createAdsReq) {
-        Ads ads = adsRepository.findAdsById(id);
+    public AdsRecord updateAds(Integer id, CreateAdsReq createAdsReq) {
+        Ads ads = adsRepository.findById(id).orElseThrow(() -> new AdsNotFoundException());
         Ads newAds = recordMapper.toEntity(createAdsReq);
         ads.setDescription(newAds.getDescription());
         ads.setPrice(newAds.getPrice());
         ads.setTitle(newAds.getTitle());
         adsRepository.save(ads);
-        return new ResponseEntity<>(recordMapper.toRecord(ads), HttpStatus.OK);
+        return recordMapper.toRecord(ads);
     }
 
-    public ResponseEntity<CommentRecord> getAdsComments(Integer adPk, Integer id) {
-        Ads ads = adsRepository.findAdsById(adPk);
+    public CommentRecord getAdsComment(Integer adPk, Integer id) {
+        Ads ads = adsRepository.findById(adPk).orElseThrow(() -> new AdsNotFoundException());
         Comment comment = commentRepository.findCommentByAdsAndId(ads, id);
-        return new ResponseEntity<>(recordMapper.toRecord(comment), HttpStatus.OK);
+        return recordMapper.toRecord(comment);
     }
 
-    public boolean deleteComments(Integer adPk, Integer id) {
-        Ads ads = adsRepository.findAdsById(adPk);
+    public boolean deleteComment(Integer adPk, Integer id) {
+        Ads ads = adsRepository.findById(adPk).orElseThrow(() -> new AdsNotFoundException());
         return commentRepository.deleteByAdsAndId(ads, id);
     }
 
-    public ResponseEntity<CommentRecord> updateComments(Integer adPk, Integer id, CommentRecord commentRecord) {
-        Ads ads = adsRepository.findAdsById(adPk);
+    public CommentRecord updateComment(Integer adPk, Integer id, CommentRecord commentRecord) {
+        Ads ads = adsRepository.findById(adPk).orElseThrow(() -> new AdsNotFoundException());
         Comment oldComment = commentRepository.findCommentByAdsAndId(ads, id);
-        Comment comment = recordMapper.toEntity(commentRecord);
-        oldComment.setText(comment.getText());
-        commentRepository.save(comment);
-        return new ResponseEntity<>(commentRecord, HttpStatus.OK);
+        oldComment.setText(commentRecord.getText());
+        commentRepository.save(oldComment);
+        return recordMapper.toRecord(oldComment);
     }
 
     public Pair<byte[], String> updateAdsImage(Integer idAds, MultipartFile image) {
@@ -138,15 +144,15 @@ public class AdsService {
         UserRecord userRecord = userService.getUser();
         User user = recordMapper.toEntity(userRecord);
 
-        List<Ads> listAds = adsRepository.findAllByAuthor(user);
+        List<Ads> listAds = adsRepository.findAllByAuthorId(user.getId());
         for (int i = 0; i < listAds.size(); i++) {
             list.add(i, recordMapper.toRecord(listAds.get(i)));
         }
         return new ResponseWrapperAds(list.size(), list);
     }
 
-    public Pair<byte[], String> getAdsByUniqueId(Integer uniqId){
-        Image image = imageRepository.getReferenceById(uniqId);
+    public Pair<byte[], String> getAdsById(Integer uniqId) {
+        Image image = imageRepository.findById(uniqId).orElseThrow(() -> new ImageNotFoundException());
         return imageService.getImageData(image);
     }
